+++ flexmeasures/data/schemas/generic_assets.py
@@ -2,6 +2,7 @@ from __future__ import annotations
 
 import json
 
+from flask import abort
 from marshmallow import validates, ValidationError, fields, validates_schema
 from flask_security import current_user
 from sqlalchemy import select
@@ -146,16 +147,26 @@ class GenericAssetTypeSchema(ma.SQLAlchemySchema):
 class GenericAssetIdField(MarshmallowClickMixin, fields.Int):
     """Field that deserializes to a GenericAsset and serializes back to an integer."""
 
-    @with_appcontext_if_needed()
-    def _deserialize(self, value, attr, obj, **kwargs) -> GenericAsset:
+    def __init__(self, status_if_not_found: int | None = None, *args, **kwargs):
+        self.status_if_not_found = status_if_not_found
+        super().__init__(*args, **kwargs)
+
+    def _deserialize(self, value: int, attr, obj, **kwargs) -> GenericAsset:
         """Turn a generic asset id into a GenericAsset."""
-        generic_asset = db.session.get(GenericAsset, value)
+        generic_asset: GenericAsset = db.session.execute(
+            select(GenericAsset).filter_by(id=int(value))
+        ).scalar_one_or_none()
         if generic_asset is None:
-            raise FMValidationError(f"No asset found with id {value}.")
+            message = f"No asset found with ID {value}."
+            if self.status_if_not_found == 404:
+                raise abort(404, message)
+            else:
+                raise FMValidationError(message)
+
         # lazy loading now (asset is somehow not in session after this)
         generic_asset.generic_asset_type
         return generic_asset
 
-    def _serialize(self, asset, attr, data, **kwargs):
+    def _serialize(self, asset: GenericAsset, attr, data, **kwargs) -> int:
         """Turn a GenericAsset into a generic asset id."""
         return asset.id