+++ documentation/api/change_log.rst
@@ -9,6 +9,7 @@ API change log
 v3.0-19 | 2024-06-03
 """"""""""""""""""""
 - New API endpoint `[POST] /assets/(id)/schedules/trigger <api/v3_0.html#post--api-v3_0-assets-(id)-schedules-trigger>`_ to schedule a site with multiple flexible devices.
+- Updated message for 404 Not Found on endpoints for managing assets: `/assets` (GET, POST) and `/assets/<id>` (GET, PATCH, DELETE).
 
 
 v3.0-18 | 2024-03-07